* Reworked the mining facility to use the resource density instead of fixed rates * The mining facility is now split up into multiple files for better readability
