+++ source/KerbalColonies.csproj
@@ -43,13 +43,17 @@
     <Compile Include="colonyClass.cs" />
     <Compile Include="colonyFacilities\KCCommNetFacility.cs" />
     <Compile Include="colonyFacilities\KCLaunchpadFacility.cs" />
+    <Compile Include="colonyFacilities\KCMiningFacility\KCMiningFacilityInfo.cs" />
+    <Compile Include="colonyFacilities\KCMiningFacility\KCMiningFacilityPlacementWindow.cs" />
+    <Compile Include="colonyFacilities\KCMiningFacility\KCMiningFacilityRate.cs" />
+    <Compile Include="colonyFacilities\KCMiningFacility\KCMiningFacilityWindow.cs" />
     <Compile Include="colonyFacilities\KCProductionFacility.cs" />
     <Compile Include="colonyFacilities\KCFacilityBase.cs" />
     <Compile Include="KCFacilityInfoClass.cs" />
     <Compile Include="colonyFacilities\KCHangarFacility.cs" />
     <Compile Include="colonyFacilities\KCKerbalFacilityBase.cs" />
     <Compile Include="colonyFacilities\KCCrewQuarters.cs" />
-    <Compile Include="colonyFacilities\KCMiningFacility.cs" />
+    <Compile Include="colonyFacilities\KCMiningFacility\KCMiningFacility.cs" />
     <Compile Include="colonyFacilities\KCResearchFacility.cs" />
     <Compile Include="colonyFacilities\KCStorageFacility.cs" />
     <Compile Include="colonyFacilities\KC_CAB_Facility.cs" />
